ADC support
Added functional ADC measurement system. To read voltage up to 20V, add a voltage divider circuit of a 1/4 ratio as this allows 20v to be read as 5v.

Changed the nomenclature of the measurement structure data type and update the battery measurements to floats, instead of doubles.

Updated existing comments to have their own header describing what they are for (even though their labels make it pretty obvious).