Streamline changes
changed how the pin assignment for the ADC was done. I added an int into the battery struct for this. The arduino reads the pin assignment (like, "A0) as an int  so thats why it's like that.

I also added a default case for when all the
batteries are charged. This way we aren't counting on a battery always needing charging. Probably need to make some more logic for it, but this is a start.