initial commit
This changes the way pins are assigned by adding an integer array into the battery structure. The array stores the digital write pins of the MOSFETS that are used to bypass the specified battery. This allows assignments, readings, and MOSFET changes to happen with array iteration instead of brute forcing it.